Moving naming_base to full modules
- moving gid_type to naming_base
- moving credit handling to naming
- flyby: remove implicit dependency of runtime_local on naming_base
