+++ hpx/runtime/naming/name.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2007-2019 Hartmut Kaiser
+//  Copyright (c) 2007-2020 Hartmut Kaiser
 //  Copyright (c) 2011 Bryce Lelbach
 //  Copyright (c) 2007 Richard D. Guidry Jr.
 //
@@ -36,756 +36,18 @@
 ///////////////////////////////////////////////////////////////////////////////
 // Version of id_type
 #define HPX_IDTYPE_VERSION  0x20
-#define HPX_GIDTYPE_VERSION 0x10
-
-///////////////////////////////////////////////////////////////////////////////
-namespace hpx { namespace naming
-{
-    namespace detail
-    {
-        ///////////////////////////////////////////////////////////////////////
-        // forward declaration
-        inline std::uint64_t strip_internal_bits_from_gid(
-            std::uint64_t msb) noexcept HPX_SUPER_PURE;
-
-        inline std::uint64_t strip_internal_bits_and_component_type_from_gid(
-            std::uint64_t msb) noexcept HPX_SUPER_PURE;
-
-        inline std::uint64_t strip_lock_from_gid(std::uint64_t msb) noexcept
-            HPX_SUPER_PURE;
-
-        inline std::uint64_t get_internal_bits(std::uint64_t msb) noexcept
-            HPX_SUPER_PURE;
-
-        inline std::uint64_t strip_internal_bits_and_locality_from_gid(
-                std::uint64_t msb) noexcept HPX_SUPER_PURE;
-
-        inline bool is_locked(gid_type const& gid) noexcept;
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    /// Global identifier for components across the HPX system.
-    struct HPX_EXPORT gid_type
-    {
-        // These typedefs are for Boost.ICL.
-        typedef gid_type size_type;
-        typedef gid_type difference_type;
-
-        static std::uint64_t const credit_base_mask = 0x1full;
-        static std::uint16_t const credit_shift = 24;
-
-        static std::uint64_t const credit_mask = credit_base_mask << credit_shift;
-        static std::uint64_t const was_split_mask = 0x80000000ull; //-V112
-        static std::uint64_t const has_credits_mask = 0x40000000ull; //-V112
-        static std::uint64_t const is_locked_mask = 0x20000000ull; //-V112
-
-        static std::uint64_t const locality_id_mask = 0xffffffff00000000ull;
-        static std::uint16_t const locality_id_shift = 32; //-V112
-
-        static std::uint64_t const virtual_memory_mask = 0x3fffffull;
-
-        // don't cache this id in the AGAS caches
-        static std::uint64_t const dont_cache_mask = 0x800000ull; //-V112
-
-        // the object is migratable
-        static std::uint64_t const is_migratable = 0x400000ull; //-V112
-
-        // Bit 64 is set for all dynamically assigned ids (if this is not set
-        // then the lsb corresponds to the lva of the referenced object).
-        static std::uint64_t const dynamically_assigned = 0x1ull;
-
-        // Bits 65-84 are used to store the component type (20 bits) if bit
-        // 64 is not set.
-        static std::uint64_t const component_type_base_mask = 0xfffffull;
-        static std::uint64_t const component_type_shift = 1ull;
-        static std::uint64_t const component_type_mask =
-            component_type_base_mask << component_type_shift;
-
-        static std::uint64_t const credit_bits_mask =
-            credit_mask | was_split_mask | has_credits_mask;
-        static std::uint64_t const internal_bits_mask = credit_bits_mask |
-            is_locked_mask | dont_cache_mask | is_migratable;
-        static std::uint64_t const special_bits_mask =
-            locality_id_mask | internal_bits_mask | component_type_mask;
-
-        gid_type () noexcept
-          : id_msb_(0), id_lsb_(0)
-        {}
-
-        explicit gid_type (std::uint64_t lsb_id) noexcept
-          : id_msb_(0), id_lsb_(lsb_id)
-        {}
-
-        explicit gid_type (std::uint64_t msb_id, std::uint64_t lsb_id) noexcept
-          : id_msb_(naming::detail::strip_lock_from_gid(msb_id)),
-            id_lsb_(lsb_id)
-        {
-        }
-
-        gid_type (gid_type const& rhs) noexcept
-          : id_msb_(naming::detail::strip_lock_from_gid(rhs.get_msb())),
-            id_lsb_(rhs.get_lsb())
-        {
-        }
-        gid_type (gid_type && rhs) noexcept
-          : id_msb_(naming::detail::strip_lock_from_gid(rhs.get_msb())),
-            id_lsb_(rhs.get_lsb())
-        {
-            rhs.id_lsb_ = rhs.id_msb_ = 0;
-        }
-
-        ~gid_type()
-        {
-            HPX_ASSERT(!is_locked());
-        }
-
-        gid_type& operator=(std::uint64_t lsb_id) noexcept
-        {
-            HPX_ASSERT(!is_locked());
-            id_msb_ = 0;
-            id_lsb_ = lsb_id;
-            return *this;
-        }
-
-        gid_type& operator=(gid_type const& rhs) noexcept
-        {
-            if (this != &rhs)
-            {
-                HPX_ASSERT(!is_locked());
-                id_msb_ = naming::detail::strip_lock_from_gid(rhs.get_msb());
-                id_lsb_ = rhs.get_lsb();
-            }
-            return *this;
-        }
-        gid_type& operator=(gid_type && rhs) noexcept
-        {
-            if (this != &rhs)
-            {
-                HPX_ASSERT(!is_locked());
-                id_msb_ = naming::detail::strip_lock_from_gid(rhs.get_msb());
-                id_lsb_ = rhs.get_lsb();
-
-                rhs.id_lsb_ = rhs.id_msb_ = 0;
-            }
-            return *this;
-        }
-
-        explicit operator bool() const noexcept
-        {
-            return 0 != id_lsb_ || 0 != id_msb_;
-        }
-
-        // We support increment, decrement, addition and subtraction
-        gid_type& operator++() noexcept         // pre-increment
-        {
-            *this += 1;
-            return *this;
-        }
-        gid_type operator++(int) noexcept       // post-increment
-        {
-            gid_type t(*this);
-            ++(*this);
-            return t;
-        }
-
-        gid_type& operator--() noexcept         // pre-decrement
-        {
-            *this -= 1;
-            return *this;
-        }
-        gid_type operator--(int)  noexcept      // post-decrement
-        {
-            gid_type t(*this);
-            --(*this);
-            return t;
-        }
-
-        // GID + GID
-        friend HPX_EXPORT gid_type operator+ (
-            gid_type const& lhs, gid_type const& rhs) noexcept;
-        gid_type operator+= (gid_type const& rhs) noexcept
-        { return (*this = *this + rhs); }
-
-        // GID + std::uint64_t
-        friend gid_type operator+ (gid_type const& lhs, std::uint64_t rhs) noexcept
-        { return lhs + gid_type(0, rhs); }
-        gid_type operator+= (std::uint64_t rhs) noexcept
-        { return (*this = *this + rhs); }
-
-        // GID - GID
-        friend HPX_EXPORT gid_type operator- (gid_type const& lhs,
-            gid_type const& rhs) noexcept;
-        gid_type operator-= (gid_type const& rhs) noexcept
-        { return (*this = *this - rhs); }
-
-        // GID - std::uint64_t
-        friend gid_type operator- (gid_type const& lhs, std::uint64_t rhs) noexcept
-        { return lhs - gid_type(0, rhs); }
-        gid_type operator-= (std::uint64_t rhs) noexcept
-        { return (*this = *this - rhs); }
-
-        friend gid_type operator& (gid_type const& lhs, std::uint64_t rhs) noexcept
-        {
-            return gid_type(lhs.id_msb_, lhs.id_lsb_ & rhs);
-        }
-
-        // comparison is required as well
-        friend bool operator== (gid_type const& lhs, gid_type const& rhs) noexcept
-        {
-            std::int64_t lhs_msb =
-                detail::strip_internal_bits_from_gid(lhs.id_msb_);
-            std::int64_t rhs_msb =
-                detail::strip_internal_bits_from_gid(rhs.id_msb_);
-
-            return (lhs_msb == rhs_msb) && (lhs.id_lsb_ == rhs.id_lsb_);
-        }
-        friend bool operator!= (gid_type const& lhs, gid_type const& rhs) noexcept
-        {
-            return !(lhs == rhs);
-        }
-
-        friend bool operator< (gid_type const& lhs, gid_type const& rhs) noexcept
-        {
-            std::int64_t lhs_msb =
-                detail::strip_internal_bits_from_gid(lhs.id_msb_);
-            std::int64_t rhs_msb =
-                detail::strip_internal_bits_from_gid(rhs.id_msb_);
-
-            if (lhs_msb < rhs_msb)
-            {
-                return true;
-            }
-            if (lhs_msb > rhs_msb)
-            {
-                return false;
-            }
-            return lhs.id_lsb_ < rhs.id_lsb_;
-        }
-
-        friend bool operator<= (gid_type const& lhs, gid_type const& rhs) noexcept
-        {
-            std::int64_t lhs_msb =
-                detail::strip_internal_bits_from_gid(lhs.id_msb_);
-            std::int64_t rhs_msb =
-                detail::strip_internal_bits_from_gid(rhs.id_msb_);
-
-            if (lhs_msb < rhs_msb)
-            {
-                return true;
-            }
-            if (lhs_msb > rhs_msb)
-            {
-                return false;
-            }
-            return lhs.id_lsb_ <= rhs.id_lsb_;
-        }
-
-        friend bool operator> (gid_type const& lhs, gid_type const& rhs) noexcept
-        {
-            return !(lhs <= rhs);
-        }
-
-        friend bool operator>= (gid_type const& lhs, gid_type const& rhs) noexcept
-        {
-            return !(lhs < rhs);
-        }
-
-        std::uint64_t get_msb() const noexcept
-        {
-            return id_msb_;
-        }
-        void set_msb(std::uint64_t msb) noexcept
-        {
-            id_msb_ = msb;
-        }
-        std::uint64_t get_lsb() const noexcept
-        {
-            return id_lsb_;
-        }
-        void set_lsb(std::uint64_t lsb) noexcept
-        {
-            id_lsb_ = lsb;
-        }
-        void set_lsb(void* lsb) noexcept
-        {
-            id_lsb_ = reinterpret_cast<std::uint64_t>(lsb);
-        }
-
-        std::string to_string() const;
-
-        // this type is at the same time its own mutex type
-        typedef gid_type mutex_type;
-
-        // Note: we deliberately don't register this lock with the lock
-        //       tracking to avoid false positives. We know that gid_types need
-        //       to be locked while suspension.
-        void lock()
-        {
-            HPX_ITT_SYNC_PREPARE(this);
-
-            while (!acquire_lock())
-            {
-                util::yield_while([this] { return is_locked(); },
-                    "hpx::naming::gid_type::lock");
-            }
-
-            util::register_lock(this);
-
-            HPX_ITT_SYNC_ACQUIRED(this);
-        }
-
-        bool try_lock()
-        {
-            HPX_ITT_SYNC_PREPARE(this);
-
-            if (acquire_lock())
-            {
-                HPX_ITT_SYNC_ACQUIRED(this);
-                util::register_lock(this);
-                return true;
-            }
-
-            HPX_ITT_SYNC_CANCEL(this);
-            return false;
-        }
-
-        void unlock()
-        {
-            HPX_ITT_SYNC_RELEASING(this);
-
-            relinquish_lock();
-            util::unregister_lock(this);
-
-            HPX_ITT_SYNC_RELEASED(this);
-        }
-
-        mutex_type& get_mutex() const noexcept
-        {
-            return const_cast<mutex_type&>(*this);
-        }
-
-    private:
-        friend HPX_EXPORT std::ostream& operator<<(std::ostream& os,
-            gid_type const& id);
-
-        friend class hpx::serialization::access;
-
-        void save(serialization::output_archive& ar, unsigned int version) const;
-
-        void load(serialization::input_archive& ar, unsigned int version);
-
-        HPX_SERIALIZATION_SPLIT_MEMBER()
-
-        // lock implementation
-        using spinlock_pool = util::spinlock_pool<gid_type>;
-
-        // returns whether lock has been acquired
-        bool acquire_lock()
-        {
-            std::lock_guard<hpx::util::detail::spinlock> l(
-                spinlock_pool::spinlock_for(this));
-            bool was_locked = (id_msb_ & is_locked_mask) ? true : false;
-            if (!was_locked)
-            {
-                id_msb_ |= is_locked_mask;
-                return true;
-            }
-            return false;
-        }
-
-        void relinquish_lock()
-        {
-            util::ignore_lock(this);
-            std::lock_guard<hpx::util::detail::spinlock> l(
-                spinlock_pool::spinlock_for(this));
-            util::reset_ignored(this);
-
-            id_msb_ &= ~is_locked_mask;
-        }
-
-        // this is used for assertions only, no need to acquire the lock
-        bool is_locked() const noexcept
-        {
-            return (id_msb_ & is_locked_mask) ? true : false;
-        }
-
-        friend bool detail::is_locked(gid_type const& gid) noexcept;
-
-        // actual gid
-        std::uint64_t id_msb_;
-        std::uint64_t id_lsb_;
-    };
-
-    HPX_EXPORT void decrement_refcnt(gid_type const& gid);
-}}
-
-///////////////////////////////////////////////////////////////////////////////
-// we know that we can serialize a gid as a byte sequence
-HPX_IS_BITWISE_SERIALIZABLE(hpx::naming::gid_type)
 
 namespace hpx { namespace naming
 {
     ///////////////////////////////////////////////////////////////////////////
-    //  Handle conversion to/from locality_id
-    inline gid_type get_gid_from_locality_id(
-        std::uint32_t locality_id) noexcept HPX_SUPER_PURE;
-
-    inline gid_type get_gid_from_locality_id(std::uint32_t locality_id) noexcept
-    {
-        return gid_type(
-            (std::uint64_t(locality_id)+1) << gid_type::locality_id_shift,
-            0);
-    }
-
-    inline std::uint32_t get_locality_id_from_gid(
-        std::uint64_t msb) noexcept HPX_PURE;
-
-    inline std::uint32_t get_locality_id_from_gid(std::uint64_t msb) noexcept
-    {
-        return std::uint32_t(msb >> gid_type::locality_id_shift) - 1;
-    }
-
-    inline std::uint32_t get_locality_id_from_gid(
-        gid_type const& id) noexcept HPX_PURE;
-
-    inline std::uint32_t get_locality_id_from_gid(gid_type const& id) noexcept
-    {
-        return get_locality_id_from_gid(id.get_msb());
-    }
-
-    inline gid_type get_locality_from_gid(gid_type const& id) noexcept
-    {
-        return get_gid_from_locality_id(get_locality_id_from_gid(id));
-    }
-
-    inline bool is_locality(gid_type const& gid) noexcept
-    {
-        return get_locality_from_gid(gid) == gid;
-    }
-
-    inline std::uint64_t replace_locality_id(std::uint64_t msb,
-        std::uint32_t locality_id) noexcept HPX_PURE;
-
-    inline std::uint64_t replace_locality_id(std::uint64_t msb,
-        std::uint32_t locality_id) noexcept
-    {
-        msb &= ~gid_type::locality_id_mask;
-        return msb | get_gid_from_locality_id(locality_id).get_msb();
-    }
-
-    inline gid_type replace_locality_id(gid_type const& gid,
-        std::uint32_t locality_id) noexcept HPX_PURE;
-
-    inline gid_type replace_locality_id(gid_type const& gid,
-        std::uint32_t locality_id) noexcept
-    {
-        std::uint64_t msb = gid.get_msb() & ~gid_type::locality_id_mask;
-        msb |= get_gid_from_locality_id(locality_id).get_msb();
-        return gid_type(msb, gid.get_lsb());
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    inline bool refers_to_virtual_memory(std::uint64_t msb) noexcept
-    {
-        return !(msb & gid_type::virtual_memory_mask);
-    }
-
-    inline bool refers_to_virtual_memory(gid_type const& gid) noexcept
-    {
-        return refers_to_virtual_memory(gid.get_msb());
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    inline bool refers_to_local_lva(gid_type const& gid) noexcept
-    {
-        return !(gid.get_msb() & gid_type::dynamically_assigned);
-    }
-
-    inline gid_type replace_component_type(gid_type const& gid,
-        std::uint32_t type) noexcept
-    {
-        std::uint64_t msb = gid.get_msb() & ~gid_type::component_type_mask;
-
-        HPX_ASSERT(!(msb & gid_type::dynamically_assigned));
-        msb |= ((std::uint64_t(type) << gid_type::component_type_shift) &
-                    gid_type::component_type_mask);
-        return gid_type(msb, gid.get_lsb());
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
     namespace detail
     {
-        // We store the log2(credit) in the gid_type
-        inline std::int16_t log2(std::int64_t val) noexcept
-        {
-            std::int16_t ret = -1;
-            while (val != 0)
-            {
-                val >>= 1;
-                ++ret;
-            }
-            return ret;
-        }
-
-        inline std::int64_t power2(std::int16_t log2credits) noexcept
-        {
-            HPX_ASSERT(log2credits >= 0);
-            return static_cast<std::int64_t>(1) << log2credits;
-        }
-
         ///////////////////////////////////////////////////////////////////////
-        inline bool has_credits(gid_type const& id) noexcept
-        {
-            return (id.get_msb() & gid_type::has_credits_mask) ? true : false;
-        }
-
-        inline bool gid_was_split(gid_type const& id) noexcept
-        {
-            return (id.get_msb() & gid_type::was_split_mask) ? true : false;
-        }
-
-        inline void set_credit_split_mask_for_gid(gid_type& id) noexcept
-        {
-            id.set_msb(id.get_msb() | gid_type::was_split_mask);
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        inline bool store_in_cache(gid_type const& id) noexcept
-        {
-            return (id.get_msb() & gid_type::dont_cache_mask) ? false : true;
-        }
-
-        inline void set_dont_store_in_cache(gid_type& gid) noexcept
-        {
-            gid.set_msb(gid.get_msb() | gid_type::dont_cache_mask);
-        }
-
         inline void set_dont_store_in_cache(id_type& id) noexcept
         {
             id.set_msb(id.get_msb() | gid_type::dont_cache_mask);
         }
 
-        ///////////////////////////////////////////////////////////////////////
-        inline bool is_migratable(gid_type const& id) noexcept
-        {
-            return (id.get_msb() & gid_type::is_migratable) ? true : false;
-        }
-
-        inline void set_is_migratable(gid_type& gid) noexcept
-        {
-            gid.set_msb(gid.get_msb() | gid_type::is_migratable);
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        inline std::int64_t get_credit_from_gid(
-            gid_type const& id) noexcept HPX_PURE;
-
-        inline std::int16_t get_log2credit_from_gid(gid_type const& id) noexcept
-        {
-            HPX_ASSERT(has_credits(id));
-            return std::int16_t((id.get_msb() >> gid_type::credit_shift) &
-                    gid_type::credit_base_mask);
-        }
-
-        inline std::int64_t get_credit_from_gid(gid_type const& id) noexcept
-        {
-            return has_credits(id) ? detail::power2(get_log2credit_from_gid(id)) : 0;
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        inline std::uint64_t strip_internal_bits_from_gid(
-            std::uint64_t msb) noexcept
-        {
-            return msb & ~gid_type::internal_bits_mask;
-        }
-
-        inline gid_type& strip_internal_bits_from_gid(gid_type& id) noexcept
-        {
-            id.set_msb(strip_internal_bits_from_gid(id.get_msb()));
-            return id;
-        }
-
-        inline std::uint64_t strip_internal_bits_except_dont_cache_from_gid(
-            std::uint64_t msb) noexcept
-        {
-            return msb & ~(gid_type::credit_bits_mask | gid_type::is_locked_mask);
-        }
-
-        inline gid_type& strip_internal_bits_except_dont_cache_from_gid(
-            gid_type& id) noexcept
-        {
-            id.set_msb(
-                strip_internal_bits_except_dont_cache_from_gid(id.get_msb()));
-            return id;
-        }
-
-        inline std::uint64_t strip_internal_bits_and_component_type_from_gid(
-            std::uint64_t msb) noexcept
-        {
-            return msb &
-                ~(gid_type::internal_bits_mask | gid_type::component_type_mask);
-        }
-
-        inline gid_type& strip_internal_bits_and_component_type_from_gid(
-            gid_type& id) noexcept
-        {
-            id.set_msb(
-                strip_internal_bits_and_component_type_from_gid(id.get_msb()));
-            return id;
-        }
-
-        inline std::uint64_t get_internal_bits(std::uint64_t msb) noexcept
-        {
-            return msb &
-                (gid_type::internal_bits_mask | gid_type::component_type_mask);
-        }
-
-        inline std::uint64_t strip_internal_bits_and_locality_from_gid(
-                std::uint64_t msb) noexcept
-        {
-            return msb &
-                (~gid_type::special_bits_mask | gid_type::component_type_mask);
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        inline std::uint32_t get_component_type_from_gid(
-            std::uint64_t msb) noexcept
-        {
-            HPX_ASSERT(!(msb & gid_type::dynamically_assigned));
-            return (msb >> gid_type::component_type_shift) &
-                gid_type::component_type_base_mask;
-        }
-
-        inline std::uint64_t add_component_type_to_gid(std::uint64_t msb,
-            std::uint32_t type) noexcept
-        {
-            HPX_ASSERT(!(msb & gid_type::dynamically_assigned));
-            return (msb & ~gid_type::component_type_mask) |
-                ((std::uint64_t(type) << gid_type::component_type_shift) &
-                    gid_type::component_type_mask);
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        inline std::uint64_t strip_lock_from_gid(std::uint64_t msb) noexcept
-        {
-            return msb & ~gid_type::is_locked_mask;
-        }
-
-        inline gid_type& strip_lock_from_gid(gid_type& gid) noexcept
-        {
-            gid.set_msb(strip_lock_from_gid(gid.get_msb()));
-            return gid;
-        }
-
-        inline bool is_locked(gid_type const& gid) noexcept
-        {
-            return gid.is_locked();
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        inline gid_type get_stripped_gid(gid_type const& id) noexcept HPX_PURE;
-
-        inline gid_type get_stripped_gid(gid_type const& id) noexcept
-        {
-            std::uint64_t const msb = strip_internal_bits_from_gid(id.get_msb());
-            std::uint64_t const lsb = id.get_lsb();
-            return gid_type(msb, lsb);
-        }
-
-        inline gid_type get_stripped_gid_except_dont_cache(
-            gid_type const& id) noexcept HPX_PURE;
-
-        inline gid_type get_stripped_gid_except_dont_cache(
-            gid_type const& id) noexcept
-        {
-            std::uint64_t const msb =
-                strip_internal_bits_except_dont_cache_from_gid(id.get_msb());
-            std::uint64_t const lsb = id.get_lsb();
-            return gid_type(msb, lsb);
-        }
-
-        inline std::uint64_t strip_credits_from_gid(std::uint64_t msb) noexcept
-        {
-            return msb & ~gid_type::credit_bits_mask;
-        }
-
-        inline gid_type& strip_credits_from_gid(gid_type& id) noexcept
-        {
-            id.set_msb(strip_credits_from_gid(id.get_msb()));
-            return id;
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        inline void set_log2credit_for_gid(
-            gid_type & id, std::int16_t log2credits) noexcept
-        {
-            // credit should be a clean log2
-            HPX_ASSERT(log2credits >= 0);
-            HPX_ASSERT(0 == (log2credits & ~gid_type::credit_base_mask));
-
-            id.set_msb((id.get_msb() & ~gid_type::credit_mask) |
-                ((std::int32_t(log2credits) << gid_type::credit_shift)
-                    & gid_type::credit_mask) |
-                gid_type::has_credits_mask);
-        }
-
-        inline void set_credit_for_gid(
-            gid_type & id, std::int64_t credits) noexcept
-        {
-            if (credits != 0)
-            {
-                std::int16_t log2credits = detail::log2(credits);
-                HPX_ASSERT(detail::power2(log2credits) == credits);
-
-                set_log2credit_for_gid(id, log2credits);
-            }
-            else
-            {
-                strip_credits_from_gid(id);
-            }
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        // has side effects, can't be pure
-        HPX_EXPORT std::int64_t add_credit_to_gid(gid_type& id,
-            std::int64_t credits);
-
-        HPX_EXPORT std::int64_t remove_credit_from_gid(gid_type& id,
-            std::int64_t debit);
-
-        HPX_EXPORT std::int64_t fill_credit_for_gid(gid_type& id,
-            std::int64_t credits = std::int64_t(HPX_GLOBALCREDIT_INITIAL));
-
-        ///////////////////////////////////////////////////////////////////////
-        HPX_EXPORT gid_type move_gid(gid_type& id);
-        HPX_EXPORT gid_type move_gid_locked(std::unique_lock<gid_type::mutex_type> l,
-            gid_type& gid);
-
-        HPX_EXPORT std::int64_t replenish_credits(gid_type& id);
-        HPX_EXPORT std::int64_t replenish_credits_locked(
-            std::unique_lock<gid_type::mutex_type>& l, gid_type& id);
-
-        ///////////////////////////////////////////////////////////////////////
-        // splits the current credit of the given id and assigns half of it to
-        // the returned copy
-        HPX_EXPORT gid_type split_credits_for_gid(gid_type& id);
-        HPX_EXPORT gid_type split_credits_for_gid_locked(
-            std::unique_lock<gid_type::mutex_type>& l, gid_type& id);
-    }
-
-    HPX_EXPORT gid_type operator+(
-        gid_type const& lhs, gid_type const& rhs) noexcept;
-    HPX_EXPORT gid_type operator-(
-        gid_type const& lhs, gid_type const& rhs) noexcept;
-
-    ///////////////////////////////////////////////////////////////////////////
-    gid_type const invalid_gid = gid_type();
-
-    ///////////////////////////////////////////////////////////////////////////
-    HPX_EXPORT std::ostream& operator<<(std::ostream& os, gid_type const& id);
-
-    namespace detail
-    {
         ///////////////////////////////////////////////////////////////////////
         enum id_type_management
         {
@@ -1007,23 +269,6 @@ namespace hpx
 {
     // pull invalid id into the main namespace
     using naming::invalid_id;
-}
-
-///////////////////////////////////////////////////////////////////////////////
-namespace std
-{
-    // specialize std::hash for hpx::naming::gid_type
-    template <>
-    struct hash<hpx::naming::gid_type>
-    {
-        std::size_t operator()(::hpx::naming::gid_type const& gid) const
-        {
-            std::size_t const h1 (std::hash<std::uint64_t>()(gid.get_lsb()));
-            std::size_t const h2 (std::hash<std::uint64_t>()(
-                hpx::naming::detail::strip_internal_bits_from_gid(gid.get_msb())));
-            return h1 ^ (h2 << 1);
-        }
-    };
 }
 
 #include <hpx/config/warnings_suffix.hpp>
