+++ libs/core/threading_base/CMakeLists.txt
@@ -104,7 +104,6 @@ add_hpx_module(
     hpx_itt_notify
     hpx_logging
     hpx_memory
-    hpx_naming_base
     hpx_type_support
     ${additional_dependencies}
   CMAKE_SUBDIRS examples tests
