+++ libs/full/components_base/CMakeLists.txt
@@ -6,6 +6,10 @@
 
 cmake_minimum_required(VERSION 3.13 FATAL_ERROR)
 
+if(NOT HPX_WITH_DISTRIBUTED_RUNTIME)
+  return()
+endif()
+
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
 # Default location is $HPX_ROOT/libs/components_base/include
@@ -50,6 +54,6 @@ add_hpx_module(
   HEADERS ${components_base_headers}
   COMPAT_HEADERS ${components_base_compat_headers}
   DEPENDENCIES hpx_core
-  MODULE_DEPENDENCIES hpx_naming
+  MODULE_DEPENDENCIES hpx_naming hpx_naming_base
   CMAKE_SUBDIRS examples tests
 )
