+++ libs/full/compute/CMakeLists.txt
@@ -43,7 +43,7 @@ set(compute_sources get_host_targets.cpp host_target.cpp numa_domains.cpp)
 
 if(HPX_WITH_DISTRIBUTED_RUNTIME)
   set(compute_optional_module_dependencies
-      hpx_actions hpx_actions_base hpx_performance_counters
+      hpx_actions hpx_actions_base hpx_naming hpx_performance_counters
       hpx_async_distributed
   )
 endif()
@@ -59,12 +59,7 @@ add_hpx_module(
   COMPAT_HEADERS ${compute_compat_headers}
   DEPENDENCIES hpx_core hpx_parallelism
   MODULE_DEPENDENCIES
-    hpx_naming
-    hpx_program_options
-    hpx_resource_partitioner
-    hpx_runtime_configuration
-    hpx_runtime_local
-    hpx_threadmanager
-    ${compute_optional_module_dependencies}
+    hpx_program_options hpx_resource_partitioner hpx_runtime_configuration
+    hpx_runtime_local hpx_threadmanager ${compute_optional_module_dependencies}
   CMAKE_SUBDIRS examples tests
 )
