+++ libs/full/compute/include/hpx/compute/host/get_targets.hpp
@@ -8,7 +8,9 @@
 
 #include <hpx/config.hpp>
 #include <hpx/futures/future_fwd.hpp>
+#if defined(HPX_HAVE_DISTRIBUTED_RUNTIME)
 #include <hpx/modules/naming.hpp>
+#endif
 
 #include <vector>
 
@@ -16,6 +18,8 @@ namespace hpx { namespace compute { namespace host {
     struct HPX_EXPORT target;
 
     HPX_EXPORT std::vector<target> get_local_targets();
+#if defined(HPX_HAVE_DISTRIBUTED_RUNTIME)
     HPX_EXPORT hpx::future<std::vector<target>> get_targets(
         hpx::id_type const& locality);
+#endif
 }}}    // namespace hpx::compute::host
