+++ libs/full/naming_base/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright (c) 2019 The STE||AR-Group
+# Copyright (c) 2019-2020 The STE||AR-Group
 #
 # SPDX-License-Identifier: BSL-1.0
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -6,20 +6,27 @@
 
 cmake_minimum_required(VERSION 3.13 FATAL_ERROR)
 
+if(NOT HPX_WITH_DISTRIBUTED_RUNTIME)
+  return()
+endif()
+
 # Default location is $HPX_ROOT/libs/naming_base/include
-set(naming_base_headers hpx/modules/naming_base.hpp)
+set(naming_base_headers hpx/naming_base/gid_type.hpp
+                        hpx/naming_base/naming_base.hpp
+)
 
 set(naming_base_compat_headers)
 
-set(naming_base_sources)
+set(naming_base_sources gid_type.cpp)
 
 include(HPX_AddModule)
 add_hpx_module(
-  core naming_base
+  full naming_base
   COMPATIBILITY_HEADERS OFF
   DEPRECATION_WARNINGS
-  GLOBAL_HEADER_GEN OFF
+  GLOBAL_HEADER_GEN ON
   HEADERS ${naming_base_headers}
-  MODULE_DEPENDENCIES hpx_config
+  SOURCES ${naming_base_sources}
+  DEPENDENCIES hpx_core
   CMAKE_SUBDIRS examples tests
 )
