+++ libs/full/naming_base/README.rst
