+++ libs/full/naming_base/docs/index.rst
