+++ libs/full/naming_base/examples/CMakeLists.txt
