+++ libs/full/naming_base/tests/CMakeLists.txt
