+++ libs/full/naming_base/tests/performance/CMakeLists.txt
