+++ libs/full/naming_base/tests/regressions/CMakeLists.txt
