+++ libs/full/naming_base/tests/unit/CMakeLists.txt
@@ -1,0 +1,24 @@
+#  Copyright (c) 2020 Hartmut Kaiser
+#
+# SPDX-License-Identifier: BSL-1.0
+#  Distributed under the Boost Software License, Version 1.0. (See accompanying
+#  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+set(tests gid_type)
+
+foreach(test ${tests})
+  set(sources ${test}.cpp)
+
+  source_group("Source Files" FILES ${sources})
+
+  set(folder_name "Tests/Unit/Modules/Full/NamingBase")
+
+  add_hpx_executable(
+    ${test}_test INTERNAL_FLAGS
+    SOURCES ${sources} ${${test}_FLAGS}
+    EXCLUDE_FROM_ALL
+    FOLDER ${folder_name}
+  )
+
+  add_hpx_unit_test("modules.naming_base" ${test} ${${test}_PARAMETERS})
+endforeach()
