+++ libs/full/runtime_local/src/thread_aware_timer.cpp
@@ -9,7 +9,7 @@
 #if defined(HPX_HAVE_THREAD_AWARE_TIMER_COMPATIBILITY)
 #include <hpx/functional/bind_front.hpp>
 #include <hpx/io_service/io_service_pool.hpp>
-#include <hpx/runtime_fwd.hpp>
+#include <hpx/runtime_local/runtime_local_fwd.hpp>
 #include <hpx/util/thread_aware_timer.hpp>
 
 #include <cstdint>
